requirements.txt
# requirements.txt
flask==2.3.3
python-telegram-bot==20.4
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0